+++ youtube_download/youtube_download.py
@@ -1,13 +1,64 @@
 from pathlib import Path
 from urllib.request import urlopen
+from functools import lru_cache
 
-from pytubefix import YouTube
+from pytubefix import YouTube, Playlist
 import yt_dlp
 from gooey import Gooey, GooeyParser
 
 
+def is_playlist(url):
+    return "/playlist?list=" in url
+
+
+def is_channel(url):
+    return "/videos" in url
+
+
+@lru_cache
+def get_ranges(indexes: str) -> list[range]:
+
+    if not indexes.strip():
+        return []
+
+    groups = indexes.replace(" ", "").split(",")
+    ranges = []
+
+    for g in groups:
+        if "-" in g:
+            start, end = g.split("-")
+            ranges.append(range(int(start), int(end) + 1))
+        else:
+            ranges.append(range(int(g), int(g) + 1))
+    return ranges
+
+
+def in_indexes(idx, indexes):
+
+    ranges = get_ranges(indexes)
+    if not ranges:
+        return True
+
+    for r in ranges:
+        if idx in r:
+            return True
+    return False
+
+
 def get_title(url):
-    return YouTube(url).title
+    if is_playlist(url):
+        title = Playlist(url).title
+    elif is_channel(url):
+        with yt_dlp.YoutubeDL({'extract_flat': True}) as ydl:
+            info = ydl.extract_info(url, download=False, process=False)
+        title = info['channel']
+    else:
+        title = YouTube(url).title
+
+    for i in "<>:\"/\\|?*":
+        title = title.replace(i, " ")
+
+    return title
 
 
 def get_description(url):
@@ -18,7 +69,7 @@ def download_video(url, path):
 
     try:
         ydl_opts = {
-            'outtmpl': f'{path}/%(title)s.%(ext)s',
+            'outtmpl': f'{path}/{get_title(url)}.%(ext)s',
             'format': 'bestvideo[ext=mp4]+bestaudio[ext=m4a]/best',
             'merge_output_format': 'mp4',
         }
@@ -42,7 +93,7 @@ def download_audio(url, path):
 
     try:
         ydl_opts = {
-            'outtmpl': f'{path}/%(title)s.%(ext)s',
+            'outtmpl': f'{path}/{get_title(url)}.%(ext)s',
             'format': 'bestaudio/best',
             'noplaylist': True,
             'postprocessors': [{
@@ -67,13 +118,13 @@ def download_audio(url, path):
 
 
 def download_info(url, path):
-    file_name = get_title(url).replace("/", "") + ".txt"
+    file_name = get_title(url) + ".txt"
     with open(Path(path) / file_name, "wb") as file:
         file.write(f"{get_title(url)}\n\n{get_description(url)}".encode("utf-8"))
 
 
 def download_thumbnail(url, path):
-    file_name = get_title(url).replace("/", "") + ".jpg"
+    file_name = get_title(url) + ".jpg"
     thumbnail_url = YouTube(url).thumbnail_url
 
     with urlopen(thumbnail_url) as response:
@@ -81,7 +132,52 @@ def download_thumbnail(url, path):
             image.write(response.read())
 
 
-@Gooey(program_name="Youtube Downloader v1.1.0",
+def process_list(url, args):
+
+    playlist = Playlist(url)
+    urls = playlist.video_urls
+
+    print(f"Розраховані індекси: '{args.indexes}' {get_ranges(args.indexes)}")
+    print(f"Завантаження {len(urls)} відео з плейліста: '{get_title(url)}' \nдо папки: {args.download_dir}")
+
+    for idx, url in enumerate(urls, start=1):
+        if in_indexes(idx, args.indexes):
+            print(f"{idx}: '{get_title(url)}")
+            process_video(url, args)
+
+
+def process_all_channel_videos(url, args):
+
+    with yt_dlp.YoutubeDL({'extract_flat': True}) as ydl:
+        info = ydl.extract_info(url, download=False, process=False)
+
+    urls = [
+        f"https://www.youtube.com/watch?v={entry['id']}"
+        for entry in info['entries']
+    ]
+
+    print(f"Розраховані індекси: '{args.indexes}' {get_ranges(args.indexes)}")
+    print(f"Завантаження {len(urls)} відео з Каналу: '{get_title(url)}' \nдо папки: {args.download_dir}")
+
+    for idx, url in enumerate(urls, start=1):
+        if in_indexes(idx, args.indexes):
+            print(f"{idx}: '{get_title(url)}")
+            process_video(url, args)
+
+
+def process_video(url, args):
+
+    if args.video:
+        download_video(url, args.download_dir)
+    if args.audio:
+        download_audio(url, args.download_dir)
+    if args.thumbnail:
+        download_thumbnail(url, args.download_dir)
+    if args.description:
+        download_info(url, args.download_dir)
+
+
+@Gooey(program_name="Youtube Downloader v1.2.0",
        default_size=(500, 600),
        clear_before_run=True)
 def main():
@@ -96,19 +192,19 @@ def main():
     group.add_argument("--audio", action="store_true", help="Save as Audio")
     group.add_argument("--thumbnail", action="store_true", help="Save Thumbnail")
     group.add_argument("--description", action="store_true", help="Save Description")
+    group.add_argument("--indexes",
+                       action="store",
+                       help="Specify items indexes (for list and chanel download): 1,2,5-12")
 
     args = parser.parse_args()
 
-    print(f"Завантаження: '{get_title(args.url)}' \nдо папки: {args.download_dir}")
-
-    if args.video:
-        download_video(args.url, args.download_dir)
-    if args.audio:
-        download_audio(args.url, args.download_dir)
-    if args.thumbnail:
-        download_thumbnail(args.url, args.download_dir)
-    if args.description:
-        download_info(args.url, args.download_dir)
+    if is_playlist(args.url):
+        process_list(args.url, args)
+    elif is_channel(args.url):
+        process_all_channel_videos(args.url, args)
+    else:
+        print(f"Завантаження: '{get_title(args.url)}' \nдо папки: {args.download_dir}")
+        process_video(args.url, args)
 
     print("Завантаження завершено")