+++ youtube_download/RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+## Version 1.2.0
+- Supports download of videos from playlist (all or specified by indexes).
+- Supports download of all videos from chanel (all or specified by indexes).
+
+Bugfix:
+- Fix failure to create file due to spec symbols to appear in file name.
+
+
 ## Version 1.1.0
 - Improved quality of downloaded video using 'yt_dlp'.
 - Download and conversion of audio to mp3 using 'yt_dlp' by default.